Convert bkod values to 64 bits in server. (#384)
To prevent time values from overflowing 28 signed bits, bkod values are extended to 64 bits at runtime and in saved games.  This should fix the 6-yearly bug where time values overflow, breaking many game systems that compare times.

Times were also sent to the client as 28 bit values, causing mail and news timestamps to overflow and then displaying incorrect values.  Now all 32 bits of the protocol can be used for timestamps.

* Test of using 64 bit ints for Blakod values.

* Convert "value" union to 64 bits.

* Convert 32 bit values from bof to 64 bit.

* Fix sending number objects to clients.

* Save timer durations as 64 bit.

This allows timers with durations of > ~37 hours.

* Add version number to saved games.

* Fix some indentation.

* Correctly handle number items sent from client.

These are still 32 bits.

+++ blakserv/timer.c
@@ -176,7 +176,7 @@ int CreateTimer(int object_id,int message_id,int milliseconds)
    return t->timer_id;
 }
 
-Bool LoadTimer(int timer_id,int object_id,char *message_name,int milliseconds)
+Bool LoadTimer(int timer_id,int object_id,char *message_name,INT64 milliseconds)
 {
    object_node *o;
    timer_node *t;
